Corrected primitive->enum conversion for Faction
The "name" attribute of a FactionPrimitive is the display string for the faction name i.e. "Star Empire". An enum member can be accessed programmatically via EnumClass['member_name']. However, the member names for CardFaction are not the same as the display names i.e. "STAR" for the member name as opposed to "Star Empire" for the member value. The new implementation compares the FactionPrimitive name ("Star Empire") with the enum member's value ("Star Empire") rather than its name ("STAR").