(feat) PlayerDeck - draw one or more cards simultaneously
You can now draw a specified number of cards from the player's deck, refilling the undrawn pile from cards in the discard pile as needed. If there are cards remaining in the deck but fewer than were requested, you will receive as many cards as possible. If you attempt to draw cards while the deck is already empty, an Indexerror will be raised. Similarly, if you request an invalid number of cards (negative, fractional, etc) an IndexError will be raised.