(refactor) Replaced UUID object with just the string representation
Each Card instance has a unique identifier used to locate the card. The original implementation assigned a UUID object to each Card instance. It shortly became apparent that comparison between UUIDs only required the string representation of the UUID in this case. In light of this all UUID objects have been replaced with their string representations to ease comparison.